fix naming new line in the end
