+++ BrowserNavigator.cs
