+++ BrowserNavigatorAdapter.cs
