+++ Form1.cs
@@ -13,7 +13,7 @@ namespace UnicornsAndRainbows
     public partial class Form1 : Form
     {
         private Configuration m_configuration;
-        private IBrowserNavigatorAdapter _mBrowserNavigatorAdapter;
+        private IBrowserNavigatorAdapter m_browserNavigatorAdapter;
 
         public Form1()
         {
@@ -22,8 +22,8 @@ namespace UnicornsAndRainbows
             UserInterface userInterface = new UserInterface(c_buttonUnicorns, c_buttonRainbows, c_labelCount);
             m_configuration = new Configuration();
             m_configuration.SearchRoot = "http://www.bing.com/images/search?q=";
-            _mBrowserNavigatorAdapter = new BrowserNavigatorAdapter();
-            BrowserNavigator browser = new BrowserNavigator(m_configuration.SearchRoot, _mBrowserNavigatorAdapter);
+            m_browserNavigatorAdapter = new BrowserNavigatorAdapter();
+            BrowserNavigator browser = new BrowserNavigator(m_configuration.SearchRoot, m_browserNavigatorAdapter);
             MouseCount mouseCount = new MouseCount();
 
             UnicornsAndRainbowsController unicornsAndRainbowsController = new UnicornsAndRainbowsController(userInterface, browser, mouseCount);
