+++ UnicornsAndRainbowsController.cs
@@ -9,12 +9,12 @@ namespace UnicornsAndRainbows
     public class UnicornsAndRainbowsController
     {
         private IUserInterface m_userInterface;
-        private IBrowserNavigator _mBrowserNavigator;
+        private IBrowserNavigator m_browserNavigator;
         private IMouseCount m_mouseCount;
 
         public UnicornsAndRainbowsController(IUserInterface userInterface, IBrowserNavigator browserNavigator, IMouseCount mouseCount)
         {
-            _mBrowserNavigator = browserNavigator;
+            m_browserNavigator = browserNavigator;
             m_userInterface = userInterface;
             m_mouseCount = mouseCount;
 
@@ -27,7 +27,7 @@ namespace UnicornsAndRainbows
 
             m_userInterface.SetCount(m_mouseCount.GetCount());
 
-            _mBrowserNavigator.NavigateTo(searchTerm);
+            m_browserNavigator.NavigateTo(searchTerm);
 
         }
     }
