+++ UnicornsAndRainbowsTests/UnicornsAndRainbowsTests.csproj
@@ -51,7 +51,7 @@
   </Choose>
   <ItemGroup>
     <Compile Include="BrowserNavigatorSimulator.cs" />
-    <Compile Include="Manager_Tests.cs" />
+    <Compile Include="Controller_Tests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="UserInterfaceSimulator.cs" />
   </ItemGroup>
